[#1087] Renamed BulkActionPanel folder.
