+++ frontend/src/bulk-action-panel/BulkActionPanel.less
