+++ frontend/src/bulk-action-panel/BulkActionPanel.tsx
