+++ frontend/src/bulk-action-panel/ClearSelection.tsx
